% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{sr_check_yml}
\alias{sr_check_yml}
\title{Utility function housing all yml map checks, passes through the map as a list object if all tests pass}
\usage{
sr_check_yml(yml_map)
}
\arguments{
\item{yml_map}{path to a yml_map file}
}
\value{
an error if any tests fail, otherwise a list object of the file mappings
}
\description{
Utility function housing all yml map checks, passes through the map as a list object if all tests pass
}
\keyword{internal}
