% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{sr_check_yml_inputs_exists}
\alias{sr_check_yml_inputs_exists}
\title{Tests that output columns are all unique}
\usage{
sr_check_yml_inputs_exists(map_list)
}
\arguments{
\item{map_list}{the list object created from reading in a yaml file}
}
\value{
an error if there is no inputs level at the top of the yaml
}
\description{
Tests that output columns are all unique
}
\keyword{internal}
