% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{sr_check_yml_io_match}
\alias{sr_check_yml_io_match}
\title{Tests that columns present in the input mappings are all present in the output columns}
\usage{
sr_check_yml_io_match(potential_inputs, output_columns)
}
\arguments{
\item{potential_inputs}{the list of potential input sources from the yaml}

\item{output_columns}{a vector of output columns from the yaml}
}
\value{
an error if there are designated output values in the inputs that are not present in the output columns
}
\description{
Tests that columns present in the input mappings are all present in the output columns
}
