% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml_mapping.R
\name{sr_check_yml_output_exists}
\alias{sr_check_yml_output_exists}
\title{Tests that there is an output level in the .yml map, and that there is a columns parameter under it}
\usage{
sr_check_yml_output_exists(map_list)
}
\arguments{
\item{map_list}{the list object created from reading in a yaml file}
}
\value{
an error if output does not exist
}
\description{
Tests that there is an output level in the .yml map, and that there is a columns parameter under it
}
\examples{
sr_check_yml_output_exists(map_list)

}
