% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{sr_check_yml_unique_output_cols}
\alias{sr_check_yml_unique_output_cols}
\title{Tests that output columns are all unique}
\usage{
sr_check_yml_unique_output_cols(output_columns)
}
\arguments{
\item{output_columns}{}
}
\value{
an error if output columns are not unique
}
\description{
Tests that output columns are all unique
}
\examples{
sr_check_yml_unique_output_cols(output_columns)

}
