% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{sr_clean_colnames}
\alias{sr_clean_colnames}
\title{Utility function for keeping column names consistent in terms of white space and case}
\usage{
sr_clean_colnames(colnames)
}
\arguments{
\item{colnames}{a vector of column names}
}
\value{
a vector of uniform column names with no leading/trailing whitespace and all lowercase
}
\description{
Utility function for keeping column names consistent in terms of white space and case
}
\keyword{internal}
