% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{sr_cleanup}
\alias{sr_cleanup}
\title{User function for cleanup}
\usage{
sr_cleanup(matched_files_list, mapping)
}
\arguments{
\item{matched_files_list}{the matched file list from sr_match}

\item{mapping}{the mapping dataframe in its original form}
}
\value{
a dataframe with cleaned nested dataframes ready for compilation
}
\description{
User function for cleanup
}
