% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{sr_compile}
\alias{sr_compile}
\title{User function for compilation}
\usage{
sr_compile(df, mapping)
}
\arguments{
\item{df}{the final cleaned dataframe from sr_cleanup}

\item{mapping}{the dataframe mapping from sr_mapping}
}
\value{
a dataframe that is all character
}
\description{
User function for compilation
}
\examples{
sr_compile(cleaned_df)

}
