% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml_mapping.R
\name{sr_convert_yml}
\alias{sr_convert_yml}
\title{Converts list map into a dataframe mapping}
\usage{
sr_convert_yml(map_list)
}
\arguments{
\item{map_list}{a list object of the yml map}
}
\value{
a dataframe of the file mapping
}
\description{
Converts list map into a dataframe mapping
}
\examples{
sr_convert_yml(map_list)

}
