% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mapping_template.R
\name{sr_create_mapping_template}
\alias{sr_create_mapping_template}
\title{User function for standing up a short .yml mapping template to get started}
\usage{
sr_create_mapping_template(mapping_file_name = "_mapping.yml")
}
\arguments{
\item{mapping_file_name}{the desired name of the .yml map. Defaults to "_mapping.yml", but can be called anything that ends in ".yml"}
}
\value{
a new .yml mapping template
}
\description{
User function for standing up a short .yml mapping template to get started
}
\examples{
sr_create_mapping_template()
sr_create_mapping_template("my_awesome_map.yml")

}
