% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{sr_find_header_row}
\alias{sr_find_header_row}
\title{Utility function for finding the row in a dataframe where column names from an input source are. (Sometimes meta data lives at the top of csv reports. We want to get rid of that)}
\usage{
sr_find_header_row(df, source_colnames, check_length = 100)
}
\arguments{
\item{df}{a dataframe we want to find the header row of}

\item{source_colnames}{a vector of column names we wish to evaluate the dataframe against}

\item{check_length}{in the event df's colnames are not the desired source_colnaes, the number of rows under the column headers to check for the column names}
}
\value{
an integer with the row number of the correct column names - 0 if the column names are correct already, NA if the column names were not found
}
\description{
Utility function for finding the row in a dataframe where column names from an input source are. (Sometimes meta data lives at the top of csv reports. We want to get rid of that)
}
