% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_matching.R
\name{sr_handle_matched_files}
\alias{sr_handle_matched_files}
\title{Takes full matched and unmatched files then reduces them down to their top matches / separates unmatched files}
\usage{
sr_handle_matched_files(matched_df)
}
\arguments{
\item{matched_df}{evaluated nested data with sources attached}
}
\value{
a list with two items: one for matched files (matched_files) and one for unmatched files (unmatched_files)
}
\description{
Takes full matched and unmatched files then reduces them down to their top matches / separates unmatched files
}
\examples{
sr_handle_matched_files(matched_df)

}
