% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{sr_list_files}
\alias{sr_list_files}
\title{Creates a tibble of file paths, preferably for csv files}
\usage{
sr_list_files(path = ".", type = "csv")
}
\arguments{
\item{path}{Relative path of file}

\item{type}{file type, e.g. "csv"}
}
\value{
a tibble of file names
}
\description{
Creates a tibble of file paths, preferably for csv files
}
