% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{sr_make_character_frames}
\alias{sr_make_character_frames}
\title{User function for cleanup}
\usage{
sr_make_character_frames(df)
}
\arguments{
\item{df}{a dataframe in need of characterized uniformity}
}
\value{
a dataframe that is all character
}
\description{
User function for cleanup
}
\keyword{internal}
