% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{sr_make_input_colnames}
\alias{sr_make_input_colnames}
\title{Utility function for putting inputs' column names in to a place they can be evaluated against nested raw data}
\usage{
sr_make_input_colnames(file_mapping)
}
\arguments{
\item{file_mapping}{a dataframe of our file mapping}
}
\value{
a nested dateframe where each row is a source and each row has clean vectorized column names
}
\description{
Utility function for putting inputs' column names in to a place they can be evaluated against nested raw data
}
\examples{
sr_make_input_colnames(file_mapping)

}
