% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_matching.R
\name{sr_match}
\alias{sr_match}
\title{User function for taking nested raw data and file mapping to identify matched and unmatched files}
\usage{
sr_match(nested_raw_data, file_mapping)
}
\arguments{
\item{nested_raw_data}{that comes from sr_import_data}

\item{file_mapping}{file map that comes from sr_yml_map}
}
\value{
a list with two items: one for matched files (matched_files) and one for unmatched files (unmatched_files)
}
\description{
User function for taking nested raw data and file mapping to identify matched and unmatched files
}
\examples{
sr_match(nested_raw_data, file_mapping)

}
