% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{sr_match_sources}
\alias{sr_match_sources}
\title{Takes raw nested data and nested input mapping to identify header rows / which files could be matched}
\usage{
sr_match_sources(df_nested_imports, input_colnames_df)
}
\arguments{
\item{df_nested_imports}{dataframe of file names and nested data}

\item{input_colnames_df}{dataframe of input sources and nested desired colnames}
}
\value{
a crossing table with all input sources and which ones matched to each file. WARNING: FILES CAN MATCH TO MULTIPLE SOURCES AT THIS STAGE. DO NOT USE THIS FUNCTION AS AN END USER.
}
\description{
Takes raw nested data and nested input mapping to identify header rows / which files could be matched
}
\examples{
sr_match_sources(df_nested_imports , input_colnames_df)

}
