% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{sr_nested_import}
\alias{sr_nested_import}
\title{Imports files into nested dataframes}
\usage{
sr_nested_import(df, path_column = NULL)
}
\arguments{
\item{df}{a dataframe with file paths}

\item{path_column}{the name of the path column}
}
\value{
a tibble of file names
}
\description{
Imports files into nested dataframes
}
