% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{sr_rename_columns}
\alias{sr_rename_columns}
\title{Utility function for renaming columns to output column names and reducing dataframes down to the desired columns in the mapping}
\usage{
sr_rename_columns(df, source, mapping)
}
\arguments{
\item{df}{a trimmed dataframe - colnames are the actual colnames}

\item{source}{the determined input source of the dataframe}

\item{mapping}{the mapping dataframe in its original form}
}
\value{
a dataframe with the desired columns
}
\description{
Utility function for renaming columns to output column names and reducing dataframes down to the desired columns in the mapping
}
\keyword{internal}
