% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch.R
\name{sr_stitch}
\alias{sr_stitch}
\title{Single function for stitching together data from disparate sources}
\usage{
sr_stitch(file_path, mapping_path, file_type = "csv",
  with_unmatched = FALSE)
}
\arguments{
\item{file_path}{path to the directory of files}

\item{mapping_path}{path to the .yml mapping file}
}
\value{
a compiled dataframe of all files that were able to be successfully matched
}
\description{
Single function for stitching together data from disparate sources
}
\examples{
sr_stitch("my/csvs/path", "_mapping.yml")
sr_stitch("my/csvs/path", "_mapping.yml", with_unmatched = TRUE)

}
