% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{sr_trim_headers}
\alias{sr_trim_headers}
\title{Utility function for getting proper column names of the original file in the column names of the dataframe, as well as trimming an excess space between the headers and the start of the data.}
\usage{
sr_trim_headers(df, row_number)
}
\arguments{
\item{df}{a dataframe}

\item{row_number}{an integer specifying the row of the dataframe with proper column headers}
}
\value{
a vector of uniform column names with no leading/trailing whitespace and all lowercase
}
\description{
Utility function for getting proper column names of the original file in the column names of the dataframe, as well as trimming an excess space between the headers and the start of the data.
}
\examples{
sr_trim_headers(my_df, 4)

}
