% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yml_mapping.R
\name{sr_yml_map}
\alias{sr_yml_map}
\title{User function for reading in a yml mapping and concerting to a dataframe}
\usage{
sr_yml_map(yml_map)
}
\arguments{
\item{yml_map}{path to a yml_map file}
}
\value{
A dataframe version of the file mapping
}
\description{
User function for reading in a yml mapping and concerting to a dataframe
}
\examples{
sr_check_yml("my_yml_map.yml")

}
