% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_docs.R
\docType{package}
\name{stitchr}
\alias{stitchr}
\alias{stitchr-package}
\title{\code{stitchr} package}
\description{
For stitching together files from disparate sources
}
\details{
See the README on
or \href{https://github.com/chriscardillo/stitchr#readme}{GitHub}
}
