% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{sr_import}
\alias{sr_import}
\title{Imports files from a directory into nested dataframes and includes file names}
\usage{
sr_import(path = ".", type = "csv")
}
\arguments{
\item{df}{a dataframe with file paths}
}
\value{
a tibble file names and nested dataframes
}
\description{
Imports files from a directory into nested dataframes and includes file names
}
\examples{
sr_import("path/to/data")

}
